+++ vaultaire-example-workflow/src/test/kotlin/com/github/manosbatsis/vaultaire/example/workflow/MagazineMockTests.kt
@@ -71,7 +71,7 @@ import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
 
-@Suppress("DEPRECATION")
+@Suppress(names = ["DEPRECATION"])
 @TestInstance(TestInstance.Lifecycle.PER_CLASS) // allow non-static @BeforeAll etc.
 class MagazineMockTests {
     companion object {